Flytt feilvisningstabell ut i egen komponent
